+++ app/Http/Controllers/TaskController.php
@@ -64,6 +64,7 @@ class TaskController extends Controller
         try {
             $validated = $request->validate([
                 'name' => 'required|string|max:255',
+                'type' => 'required|string',
                 'project_id' => 'required|exists:projects,id',
                 'assigned' => 'nullable|exists:users,id',
                 'priority' => 'nullable|string',
@@ -90,15 +91,14 @@ class TaskController extends Controller
                     "value" => "  "
                 ];
             }
-
-
-
-
             $project = Project::find($request->project_id);
+            $maxOrder = Task::max('order_column');
 
+            // Task::create($validated);
 
-            Task::create($validated);
-
+            $newTask = Task::create(array_merge($validated, [
+                'order_column' => $maxOrder ? $maxOrder + 1 : 0,  // If no tasks exist, set it to 0
+            ]));
 
             $project->activities()->create([
                 'user_id' => Auth::id(),
@@ -148,9 +148,16 @@ class TaskController extends Controller
     public function update(Request $request, Task $task)
     {
         // dd($request);
+        $project = $task->project()->first();
+        $members = ProjectMember::where("user_id", auth()->user()->id)->where('project_id', $project->id)->first();
+
+        if ($members->role === "member" && $task->assigned !== auth()->user()->id) {
+            return back()->withErrors(['Error' => 'you are not authorized to change']);
+        }
         try {
             $validated = $request->validate([
                 'name' => 'required|string|max:255',
+                'type' => 'required|string',
                 'assigned' => 'nullable|exists:users,id',
                 'status' => 'string',
                 'approved' => 'nullable',
@@ -163,18 +170,18 @@ class TaskController extends Controller
 
             $task->update([
                 'name' => $validated['name'],
-                'assigned' => $validated['assigned'] ?? $task->assigned,
+                'type' => $validated['type'],
+                'assigned' => $validated['assigned'],
                 'status' => $validated['status'] ?? $task->status,
                 'approved' => $validated['approved'] ?? $task->approved,
-                'priority' => $validated['priority'] ?? $task->priority,
-                'due_date' => $validated['due_date'] ?? $task->due_date,
+                'priority' => $validated['priority'],
+                'due_date' => $validated['due_date'],
                 'additional_column' => $validated['additional_column'] ?? $task->additional_column,
                 'updated_by' => auth()->id(),
                 'description' => $validated['description'] ?? $task->description, // add this
                 'parent_task_id' => $validated['parent_task_id'] ?? $task->parent_task_id
             ]);
 
-            // dd('Hello2');
 
             $project = Project::find($task->project_id);
 
@@ -183,7 +190,7 @@ class TaskController extends Controller
                 'activity' => 'Update Task called ' . $request->name,
             ]);
         } catch (Exception $ex) {
-            dd($ex->getMessage);
+            dd($ex);
         }
         // return redirect()->route('task.index')->with('success', 'Task updated successfully.');
     }
@@ -203,17 +210,21 @@ class TaskController extends Controller
      * Remove the specified resource from storage.
      */
     public function destroy(Task $task)
-
     {
         try {
-            $project = Project::find($task->project_id);
+            $project = $task->project()->first();
+            $members = ProjectMember::where("user_id", auth()->user()->id)->where('project_id', $project->id)->first();
+
+            if ($members->role === "member" && $task->assigned !== auth()->user()->id) {
+                return back()->withErrors(['Error' => 'you are not authorized to change']);
+            }
 
             $project->activities()->create([
                 'user_id' => Auth::id(),
                 'activity' => ' Deleted Task called ' . $task->name,
             ]);
             $task->delete();
-            return redirect()->route('task.index')->with('success', 'Task deleted successfully.');
+            return redirect()->route('project.show', $task->project_id)->with('success', 'Task deleted successfully.');
         } catch (Exception $ex) {
             dd($ex);
         }
